+++ History.markdown
@@ -5,11 +5,14 @@
   * Replace `puts` calls with `Jekyll.logger.info` (#69)
   * Formatting of dates and times in a DRY manner (#60)
   * Appease newest Rubocop (#71)
+  * chore(deps): rubocop-jekyll 0.3 (#80)
 
 ### Minor Enhancements
 
   * Add some color to the success msg like jekyll new (#75)
   * Allow additional front matter for Post (#41)
+  * Mirror `draft` command to current `post` command (#79)
+  * Generate configuration from CLI options (#76)
 
 ## 0.8.0 / 2018-03-24